Move profiling functions into RDT library
