+++ rdt/performance/__init__.py
@@ -1,0 +1,9 @@
+"""Functions to evaluate and test the performance of the RDT Transformers."""
+
+from rdt.performance import profiling
+from rdt.performance.performance import evaluate_transformer_performance
+
+__all__ = [
+    'evaluate_transformer_performance',
+    'profiling',
+]
