+++ rdt/performance/datasets/datetime.py
@@ -6,8 +6,8 @@ from abc import ABC
 import numpy as np
 import pandas as pd
 
-from tests.datasets.base import BaseDatasetGenerator
-from tests.datasets.utils import add_nans
+from rdt.performance.datasets.base import BaseDatasetGenerator
+from rdt.performance.datasets.utils import add_nans
 
 
 class DatetimeGenerator(BaseDatasetGenerator, ABC):
