+++ tests/contributing.py
@@ -12,13 +12,14 @@ import pytest
 from tabulate import tabulate
 
 from rdt.transformers import get_transformer_class, get_transformers_by_type
+from rdt.performance import evaluate_transformer_performance
+from rdt.performance.datasets import get_dataset_generators_by_type
 from tests.code_style import (
     get_test_location, validate_test_location, validate_test_names, validate_transformer_addon,
     validate_transformer_importable_from_parent_module, validate_transformer_module,
     validate_transformer_name, validate_transformer_subclass)
-from tests.datasets import get_dataset_generators_by_type
 from tests.integration.test_transformers import validate_transformer
-from tests.performance import evaluate_transformer_performance, validate_performance
+from tests.performance import validate_performance
 from tests.quality.test_quality import (
     TEST_THRESHOLD, get_regression_scores, get_results_table, get_test_cases)
