+++ tests/performance/__init__.py
@@ -1,12 +1,10 @@
 """Functions to evaluate and test the performance of the RDT Transformers."""
 
 from tests.performance import profiling
-from tests.performance.test_performance import (
-    TEST_NAMES, evaluate_transformer_performance, validate_performance)
+from tests.performance.test_performance import evaluate_transformer_performance, validate_performance
 
 __all__ = [
     'evaluate_transformer_performance',
     'profiling',
-    'TEST_NAMES',
     'validate_performance',
 ]
